[Cleanup] Extract shared scheduler test dummies into tests/utils.py
