+++ tests/test_scheduler.py
@@ -8,35 +8,15 @@ import pytest
 from sglang_omni.engines.omni.scheduler import Scheduler
 from sglang_omni.engines.omni.types import SchedulerStatus
 
-
-class _DummyBatchPlanner:
-    def select_requests(self, *args, **kwargs):
-        return []
-
-    def build_batch(self, *args, **kwargs):
-        return None
-
-
-class _DummyResourceManager:
-    def free(self, request):
-        del request
-
-
-class _DummyIterationController:
-    def update_request(self, request, output):
-        del request, output
-
-    def is_finished(self, request, output):
-        del request, output
-        return False
+from tests.utils import DummyBatchPlanner, DummyIterationController, DummyResourceManager
 
 
 @pytest.mark.asyncio
 async def test_get_result_keeps_terminal_request_streamable() -> None:
     scheduler = Scheduler(
-        _DummyBatchPlanner(),
-        _DummyResourceManager(),
-        _DummyIterationController(),
+        DummyBatchPlanner(),
+        DummyResourceManager(),
+        DummyIterationController(),
     )
     scheduler.add_request("req-1", {"value": 1})
     request = scheduler.requests["req-1"]
