+++ tests/utils.py
@@ -11,3 +11,33 @@ def find_free_port() -> int:
     with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
         s.bind(("", 0))
         return s.getsockname()[1]
+
+
+class DummyBatchPlanner:
+    """No-op batch planner for scheduler unit tests."""
+
+    def select_requests(self, waiting_reqs, running_reqs, resource_manager):
+        del waiting_reqs, running_reqs, resource_manager
+        return []
+
+    def build_batch(self, selected):
+        del selected
+        return None
+
+
+class DummyResourceManager:
+    """No-op resource manager for scheduler unit tests."""
+
+    def free(self, request):
+        del request
+
+
+class DummyIterationController:
+    """No-op iteration controller for scheduler unit tests."""
+
+    def update_request(self, request, output):
+        del request, output
+
+    def is_finished(self, request, output):
+        del request, output
+        return False
